wip(task-2): extract RoleSelect, RolePicker gate green
