+++ src/components/roles/RolePicker.tsx
@@ -1,29 +1,14 @@
-import { useMemo, useState } from 'react';
+import { useState } from 'react';
 import { Button } from '@/components/ui/button';
-import {
-  Command,
-  CommandGroup,
-  CommandInput,
-  CommandItem,
-  CommandList,
-} from '@/components/ui/command';
-import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
-import { useInsideScrollLock } from '@/components/ui/scroll-lock-boundary';
-import { Check, ChevronsUpDown, ArrowUp, ArrowDown } from 'lucide-react';
+import { ArrowUp, ArrowDown } from 'lucide-react';
 import { useRoles, type RoleWithGrants } from '@/hooks/useRoles';
 import { useAssignRole, assignRoleErrorMessage } from '@/hooks/useAssignRole';
 import { useToast } from '@/hooks/use-toast';
-import { RoleAreaChips } from '@/components/roles/RoleAreaChips';
+import { RoleSelect } from '@/components/roles/RoleSelect';
 import { ROLE_METADATA } from '@/lib/permissions/definitions';
-import {
-  CUSTOM_ROLE,
-  canInviteCustomRole,
-  getInvitableRoles,
-  isAssignableCustomRole,
-} from '@/lib/permissions/invitations';
+import { CUSTOM_ROLE } from '@/lib/permissions/invitations';
 import { roleDelta, type RoleGrantSet } from '@/lib/permissions/roleDelta';
 import type { Role } from '@/lib/permissions/types';
-import { cn } from '@/lib/utils';
 
 export interface RolePickerProps {
   membershipId: string;
@@ -73,34 +58,15 @@ export function RolePicker({
   onAssigned,
 }: RolePickerProps) {
   const [open, setOpen] = useState(false);
-  const [search, setSearch] = useState('');
   const [candidateId, setCandidateId] = useState<string | null>(null);
-  const modal = useInsideScrollLock();
   const { toast } = useToast();
 
   // useRoles returns { roles, isLoading, error, ... } — NOT a raw React Query
-  // result, so there is no `data` here.
-  const { roles, isLoading, error } = useRoles(restaurantId);
+  // result, so there is no `data` here. RolePicker still needs `roles` for
+  // currentRow/candidateRow/currentLabel; RoleSelect makes the same call
+  // under the same React Query key, so this is one fetch, two readers.
+  const { roles } = useRoles(restaurantId);
   const assign = useAssignRole(restaurantId);
-
-  const invitable = useMemo(() => getInvitableRoles(callerRole), [callerRole]);
-  const mayAssignCustom = canInviteCustomRole(callerRole);
-
-  const searchQuery = search.trim().toLowerCase();
-  const matches = (name: string) => name.toLowerCase().includes(searchQuery);
-
-  // Only roles `assign_membership_role` will actually accept for
-  // `collaborator_custom` — see `isAssignableCustomRole` for why the three
-  // checks are all needed. Shared with `canAssignTargetRole`, which gates the
-  // "Assign people" entry points, so the options here and the door that leads to
-  // them agree.
-  const customRoles = roles.filter((r) => isAssignableCustomRole(r, restaurantId) && matches(r.name));
-  const builtinRoles = roles.filter(
-    (r) =>
-      r.legacy_role !== null &&
-      (invitable as readonly string[]).includes(r.legacy_role) &&
-      matches(r.name)
-  );
 
   const isCurrent = (r: RoleWithGrants) =>
     currentRoleId ? r.id === currentRoleId : r.legacy_role === currentRole;
@@ -139,7 +105,6 @@ export function RolePicker({
           toast({ title: `${personName} is now ${candidateRow.name}` });
           setOpen(false);
           setCandidateId(null);
-          setSearch('');
           onAssigned?.();
         },
         onError: (err) =>
@@ -158,148 +123,54 @@ export function RolePicker({
   // hence the name is embedded rather than replaced.
   const triggerLabel = `${personName}: role is ${currentLabel}. Change role`;
 
-  const renderRow = (r: RoleWithGrants) => (
-    // value={r.name} is load-bearing: cmdk filters on the `value` prop, and
-    // with chips and a description as children its default text extraction
-    // would match the concatenated blob instead of the name.
-    <CommandItem
-      key={r.id}
-      value={r.name}
-      onSelect={() => setCandidateId(r.id)}
-      className="flex flex-col items-start gap-1.5 py-2.5"
-    >
-      <div className="flex w-full items-center gap-2">
-        <Check
-          className={cn('h-4 w-4 shrink-0', isCurrent(r) ? 'opacity-100' : 'opacity-0')}
-        />
-        <span className="text-[14px] font-medium text-foreground">{r.name}</span>
-      </div>
-      {r.description && (
-        <p className="pl-6 text-[13px] text-muted-foreground">{r.description}</p>
-      )}
-      <div className="pl-6">
-        <RoleAreaChips areas={r.role_areas} />
-      </div>
-    </CommandItem>
-  );
-
   return (
-    <Popover
+    <RoleSelect
+      restaurantId={restaurantId}
+      callerRole={callerRole}
+      // The CURRENT role, not the candidate — `isCurrent` above never moves
+      // when you click an option, and the footer is what shows the pending
+      // choice. Passing `candidateId` here would erase the only on-screen
+      // record of the role the person holds today.
+      value={currentRow?.id ?? null}
+      onSelect={(r) => setCandidateId(r.id)}
+      triggerLabel={triggerLabel}
+      triggerText={currentLabel}
+      disabled={disabled}
       open={open}
       onOpenChange={(next) => {
         setOpen(next);
-        if (!next) {
-          setCandidateId(null);
-          setSearch('');
-        }
+        if (!next) setCandidateId(null);
       }}
-      modal={modal}
-    >
-      <PopoverTrigger asChild>
-        <Button
-          variant="outline"
-          role="combobox"
-          aria-expanded={open}
-          aria-label={triggerLabel}
-          disabled={disabled}
-          className="h-7 max-w-[220px] gap-1 rounded-full border-border/40 px-2.5 text-[13px] font-medium"
-        >
-          <span className="truncate">{currentLabel}</span>
-          <ChevronsUpDown className="h-3 w-3 shrink-0 opacity-50" />
-        </Button>
-      </PopoverTrigger>
-
-      <PopoverContent className="w-[340px] p-0" align="end">
-        <Command shouldFilter={false}>
-          <CommandInput
-            placeholder="Search roles..."
-            value={search}
-            onValueChange={setSearch}
-          />
-          <CommandList>
-            {/*
-              These three states are direct children of CommandList, never
-              routed through CommandEmpty. CommandEmpty means "no rows
-              registered" — it cannot distinguish a load failure from an empty
-              restaurant, and rendering an error through it would tell the
-              admin there are no roles when in fact the request failed.
-
-              !restaurantId is checked alongside isLoading because useRoles is
-              enabled: !!restaurantId, and a disabled React Query reports
-              isLoading === false.
-            */}
-            {(!restaurantId || isLoading) && (
-              <p
-                role="status"
-                aria-live="polite"
-                className="px-3 py-6 text-center text-[13px] text-muted-foreground"
-              >
-                Loading roles…
+      footer={
+        delta && candidateRow ? (
+          <div className="space-y-2 border-t border-border/40 px-3 py-3">
+            {delta.isSame ? (
+              <p className="text-[13px] text-muted-foreground">
+                Same access — only the label changes.
               </p>
-            )}
-            {error && (
-              <p role="alert" className="px-3 py-6 text-center text-[13px] text-destructive">
-                Couldn't load roles. Please try again.
-              </p>
-            )}
-            {!isLoading && !error && customRoles.length === 0 && builtinRoles.length === 0 && (
-              <p className="px-3 py-6 text-center text-[13px] text-muted-foreground">
-                No roles match. Create one in Team → Roles.
-              </p>
-            )}
-
-            {/*
-              The custom group renders for a caller who cannot assign custom
-              roles too — visibly, with the reason. Hiding it is what produced
-              the original "I made a role and it's nowhere" confusion.
-            */}
-            {customRoles.length > 0 && (
-              <CommandGroup heading="Your custom roles">
-                {mayAssignCustom ? (
-                  customRoles.map(renderRow)
-                ) : (
-                  <p className="px-2 py-2 text-[13px] text-muted-foreground">
-                    Only an owner or manager can assign a custom role.
-                  </p>
+            ) : (
+              <div className="space-y-1">
+                {renderDeltaLines(
+                  [...delta.gains.map((g) => g.label), ...delta.flagGains.map((f) => f.label)],
+                  'gain'
                 )}
-              </CommandGroup>
+                {renderDeltaLines(
+                  [...delta.loses.map((l) => l.label), ...delta.flagLoses.map((f) => f.label)],
+                  'lose'
+                )}
+              </div>
             )}
-
-            {builtinRoles.length > 0 && (
-              <CommandGroup heading="Built-in">{builtinRoles.map(renderRow)}</CommandGroup>
-            )}
-          </CommandList>
-
-          {delta && candidateRow && (
-            <div className="space-y-2 border-t border-border/40 px-3 py-3">
-              {delta.isSame ? (
-                <p className="text-[13px] text-muted-foreground">
-                  Same access — only the label changes.
-                </p>
-              ) : (
-                <div className="space-y-1">
-                  {renderDeltaLines(
-                    [...delta.gains.map((g) => g.label), ...delta.flagGains.map((f) => f.label)],
-                    'gain'
-                  )}
-                  {renderDeltaLines(
-                    [...delta.loses.map((l) => l.label), ...delta.flagLoses.map((f) => f.label)],
-                    'lose'
-                  )}
-                </div>
-              )}
-              <Button
-                onClick={commit}
-                disabled={assign.isPending}
-                aria-label={`Change role to ${candidateRow.name}`}
-                className="h-9 w-full rounded-lg bg-foreground text-[13px] font-medium text-background hover:bg-foreground/90"
-              >
-                {assign.isPending ? 'Changing…' : `Change role to ${candidateRow.name}`}
-              </Button>
-            </div>
-          )}
-        </Command>
-      </PopoverContent>
-    </Popover>
+            <Button
+              onClick={commit}
+              disabled={assign.isPending}
+              aria-label={`Change role to ${candidateRow.name}`}
+              className="h-9 w-full rounded-lg bg-foreground text-[13px] font-medium text-background hover:bg-foreground/90"
+            >
+              {assign.isPending ? 'Changing…' : `Change role to ${candidateRow.name}`}
+            </Button>
+          </div>
+        ) : null
+      }
+    />
   );
 }
